Avoid database rollback when upload contains duplicate keys
By finding and filtering duplicates (conflicts) before inserting them
into the database, we avoid triggering a rollback of the transaction.
Additionally, we need to tell the @Transaction not to rollback on
`DiagnosisKeyDuplicateInsertException`'s, which are triggered when
duplicates were detected.

In case of exceptions while trying to INSERT a key into the database,
we have to assume a rollback on database level will be triggered and
therefore report all keys as failed (empty 201 list).